# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Souklick Review Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user and organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /organizations/me:
    get:
      operationId: getMyOrganization
      tags: [organizations]
      summary: Get the current user's organization
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"

    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization settings including brand voice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /locations:
    get:
      operationId: getLocations
      tags: [locations]
      summary: Get all locations for the organization
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationWithStats"

    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a new location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationRequest"
      responses:
        "201":
          description: Created location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      summary: Get a specific location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Location details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationWithStats"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Update a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationRequest"
      responses:
        "200":
          description: Updated location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"

    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /ai/generate-response:
    post:
      operationId: generateAiResponse
      tags: [ai]
      summary: Generate an AI response draft for a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateResponseRequest"
      responses:
        "200":
          description: AI-generated response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedResponse"
        "500":
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [owner, manager, staff]
        organizationId:
          type: string
        notificationEmail:
          type: boolean
        notificationPush:
          type: boolean
        notificationMinRating:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - fullName
        - role
        - organizationId
        - notificationEmail
        - notificationPush
        - notificationMinRating
        - createdAt
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        organizationName:
          type: string
      required:
        - email
        - password
        - fullName
        - organizationName
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        organization:
          $ref: "#/components/schemas/Organization"
      required:
        - user
        - organization
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        brandVoiceFormality:
          type: string
          enum: [casual, balanced, professional]
        brandVoiceEmojis:
          type: string
          enum: [never, sometimes, often]
        brandVoiceSignoff:
          type: string
          nullable: true
        brandVoiceExamples:
          type: array
          items:
            type: string
          nullable: true
        subscriptionPlan:
          type: string
          enum: [trial, starter, growth, enterprise]
        subscriptionStatus:
          type: string
          enum: [active, past_due, cancelled]
        trialEndsAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - brandVoiceFormality
        - brandVoiceEmojis
        - subscriptionPlan
        - subscriptionStatus
        - createdAt
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
        brandVoiceFormality:
          type: string
          enum: [casual, balanced, professional]
        brandVoiceEmojis:
          type: string
          enum: [never, sometimes, often]
        brandVoiceSignoff:
          type: string
          nullable: true
        brandVoiceExamples:
          type: array
          items:
            type: string
    Location:
      type: object
      properties:
        id:
          type: string
        organizationId:
          type: string
        name:
          type: string
        address:
          type: string
          nullable: true
        googlePlaceId:
          type: string
          nullable: true
        zomatoRestaurantId:
          type: string
          nullable: true
        tripadvisorLocationId:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - name
        - isActive
        - createdAt
    LocationWithStats:
      allOf:
        - $ref: "#/components/schemas/Location"
        - type: object
          properties:
            averageRating:
              type: number
              nullable: true
            reviewCount:
              type: integer
            responseRate:
              type: number
          required:
            - reviewCount
            - responseRate
    CreateLocationRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        googlePlaceId:
          type: string
        zomatoRestaurantId:
          type: string
        tripadvisorLocationId:
          type: string
      required:
        - name
    UpdateLocationRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        googlePlaceId:
          type: string
        zomatoRestaurantId:
          type: string
        tripadvisorLocationId:
          type: string
        isActive:
          type: boolean
...
    GenerateResponseRequest:
      type: object
      properties:
        reviewId:
          type: string
        userNotes:
          type: string
      required:
        - reviewId
    GeneratedResponse:
      type: object
      properties:
        draftText:
          type: string
        responseId:
          type: string
      required:
        - draftText
        - responseId
...