# lib/api-zod/src/generated/types/generatedResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Souklick Review Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedResponse {
  draftText: string;
  responseId: string;
}